Fix for model connectivity issues during train and test
Fixed (1) dual ca3 -> ca1 pathways (this pathway was erroneously duplicated in ConfigNet) and (2) CA3 -> CA1 was off during test; now it is on alongside ECin -> CA1 on during test. An additional parameter change is made increasing ca1FmECin.WtScale.Abs from 1 to 3 during test. These implemented changes maintain the same performance as before. Note that this model is optimized for performance on the statistical learning task from Schapiro et al., 2017 and the ca1FmECin.WtScale.Abs change is necessary for that. The change could in theory impact performance on episodic memory tasks, which were not investigated here.